added two tcp tests. Need more fixes
