+++ nifi-nar-bundles/nifi-elasticsearch-bundle/nifi-elasticsearch-processors/src/test/java/org/apache/nifi/processors/elasticsearch/TestElasticsearchIntegration.java
@@ -14,6 +14,7 @@ import org.apache.nifi.util.TestRunners;
 import org.junit.*;
 
 import java.io.IOException;
+import java.io.InputStream;
 import java.sql.Date;
 import java.sql.Time;
 import java.sql.Timestamp;
@@ -27,8 +28,10 @@ import static org.junit.Assert.assertNotNull;
 public class TestElasticsearchIntegration extends ElasticsearchDockerInitializer {
     private TestRunner runner;
     private static byte[] docExample;
+    private InputStream docExampleStream;
     private static String esUrl;
     private static String esUrlProxy;
+    private static String esTcpPort;
     private static String proxyPort;
     private static String proxyAuthPort;
     private static Boolean dockerNetworkExistedBefore;
@@ -50,8 +53,9 @@ public class TestElasticsearchIntegration extends ElasticsearchDockerInitializer
         }
         logger.info("Elasticsearch cluster nodes ip addresses:"+ elasticsearchNodesIps);
         EnumMap<DockerServicePortType, String> elasticsearchSquidDockerServicesPorts = getElasticsearchSquidFreePorts();
-        esUrl = "http://127.0.0.1:" + elasticsearchSquidDockerServicesPorts.get(DockerServicePortType.ES01_SP);
+        esUrl = "http://127.0.0.1:" + elasticsearchSquidDockerServicesPorts.get(DockerServicePortType.ES_NIFI_01_HTTP_PORT);
         esUrlProxy = "http://" + elasticsearchServerHosts.get(ElasticsearchNodesType.ES_NODE_01_IP_ADDRESS) + ":9200";
+        esTcpPort = elasticsearchSquidDockerServicesPorts.get(DockerServicePortType.ES_NIFI_01_TCP_PORT);
         proxyPort = elasticsearchSquidDockerServicesPorts.get(DockerServicePortType.SQUID_SP);
         proxyAuthPort = elasticsearchSquidDockerServicesPorts.get(DockerServicePortType.SQUID_AUTH_SP);
         clearElasticsearchSquidDocker();
@@ -74,6 +78,8 @@ public class TestElasticsearchIntegration extends ElasticsearchDockerInitializer
     @Before
     public void once() throws IOException {
         docExample = TestPutElasticsearchHttp.getDocExample();
+        ClassLoader classloader = Thread.currentThread().getContextClassLoader();
+        docExampleStream = classloader.getResourceAsStream("DocumentExample.json");
     }
 
     @After
@@ -360,5 +366,81 @@ public class TestElasticsearchIntegration extends ElasticsearchDockerInitializer
         runner.enqueue(docExample);
         runner.run(1, true, true);
         runner.assertAllFlowFilesTransferred(FetchElasticsearchHttp.REL_SUCCESS, 1);
+    }
+
+    @Test
+    public void testPutElasticSearchBasic() {
+        System.out.println("Starting test " + new Object() {
+        }.getClass().getEnclosingMethod().getName());
+        final TestRunner runner = TestRunners.newTestRunner(new PutElasticsearch());
+
+        //Local Cluster - Mac pulled from brew
+        runner.setProperty(AbstractElasticsearchTransportClientProcessor.CLUSTER_NAME, "elasticsearch_brew");
+        runner.setProperty(AbstractElasticsearchTransportClientProcessor.HOSTS, "127.0.0.1:"+esTcpPort);
+        runner.setProperty(AbstractElasticsearchTransportClientProcessor.PING_TIMEOUT, "5s");
+        runner.setProperty(AbstractElasticsearchTransportClientProcessor.SAMPLER_INTERVAL, "5s");
+
+        runner.setProperty(PutElasticsearch.INDEX, "doc");
+        runner.setProperty(PutElasticsearch.BATCH_SIZE, "1");
+
+        runner.setProperty(PutElasticsearch.TYPE, "status");
+        runner.setProperty(PutElasticsearch.ID_ATTRIBUTE, "doc_id");
+        runner.assertValid();
+
+        runner.enqueue(docExample, new HashMap<String, String>() {{
+            put("doc_id", "28039652140");
+        }});
+
+
+        runner.enqueue(docExample);
+        runner.run(1, true, true);
+
+        runner.assertAllFlowFilesTransferred(PutElasticsearch.REL_SUCCESS, 1);
+    }
+
+    @Test
+    public void testPutElasticSearchBatch() throws IOException {
+        System.out.println("Starting test " + new Object() {
+        }.getClass().getEnclosingMethod().getName());
+        final TestRunner runner = TestRunners.newTestRunner(new PutElasticsearch());
+
+        //Local Cluster - Mac pulled from brew
+        runner.setProperty(AbstractElasticsearchTransportClientProcessor.CLUSTER_NAME, "elasticsearch_brew");
+        runner.setProperty(AbstractElasticsearchTransportClientProcessor.HOSTS, "127.0.0.1:"+esTcpPort);
+        runner.setProperty(AbstractElasticsearchTransportClientProcessor.PING_TIMEOUT, "5s");
+        runner.setProperty(AbstractElasticsearchTransportClientProcessor.SAMPLER_INTERVAL, "5s");
+        runner.setProperty(PutElasticsearch.INDEX, "doc");
+        runner.setProperty(PutElasticsearch.BATCH_SIZE, "100");
+
+        runner.setProperty(PutElasticsearch.TYPE, "status");
+        runner.setProperty(PutElasticsearch.ID_ATTRIBUTE, "doc_id");
+        runner.assertValid();
+
+
+        String message = convertStreamToString(docExampleStream);
+        for (int i = 0; i < 100; i++) {
+
+            long newId = 28039652140L + i;
+            final String newStrId = Long.toString(newId);
+            runner.enqueue(message.getBytes(), new HashMap<String, String>() {{
+                put("doc_id", newStrId);
+            }});
+
+        }
+
+        runner.run();
+
+        runner.assertAllFlowFilesTransferred(PutElasticsearch.REL_SUCCESS, 100);
+    }
+
+    /**
+     * Convert an input stream to a stream
+     *
+     * @param is input the input stream
+     * @return return the converted input stream as a string
+     */
+    static String convertStreamToString(java.io.InputStream is) {
+        java.util.Scanner s = new java.util.Scanner(is).useDelimiter("\\A");
+        return s.hasNext() ? s.next() : "";
     }
 }
