+++ nifi-nar-bundles/nifi-elasticsearch-bundle/nifi-elasticsearch-processors/src/test/java/org/apache/nifi/processors/elasticsearch/docker/DockerServicePortType.java
@@ -1,8 +1,10 @@
 package org.apache.nifi.processors.elasticsearch.docker;
 
 public enum DockerServicePortType {
-    ES01_SP,
-    ES02_SP,
+    ES_NIFI_01_HTTP_PORT,
+    ES_NIFI_02_HTTP_PORT,
+    ES_NIFI_01_TCP_PORT,
+    ES_NIFI_02_TCP_PORT,
     SQUID_SP,
     SQUID_AUTH_SP
 }
