+++ nifi-nar-bundles/nifi-elasticsearch-bundle/nifi-elasticsearch-processors/src/test/resources/es1.yml
@@ -1,4 +1,4 @@
-cluster.name: elasticsearch-cluster
+cluster.name: elasticsearch_brew
 node.name: es-node1
 http.port: 9200
 transport.tcp.port: 9300
